traduccion de componentes funcionando menos los proyectos
