+++ src/components/sections/Projects.astro
@@ -1,6 +1,6 @@
 ---
 import CardProject from "@/components/cards/CardProject.astro";
-import { PROJECTS } from '@/data/projects_detail';
+import { PROJECTS } from '@/data/projects_details';
 import CardColaboracion from "@/components/cards/cardColaboracion.astro";
 import { PROJECTSC } from "@/data/projects_colaborate";
 import { getLangFromUrl, useTranslations } from '@/i18n/utils';
@@ -8,7 +8,10 @@ import { getLangFromUrl, useTranslations } from '@/i18n/utils';
 const lang = getLangFromUrl(Astro.url);
 const t = useTranslations(lang);
 
-const isSingleProject = PROJECTS.length === 1;
+const projects = PROJECTS[lang as keyof typeof PROJECTS] || [];
+const collaborationProjects = PROJECTSC[lang as keyof typeof PROJECTSC] || [];
+
+const isSingleProject = projects.length === 1;
 ---
 
 <section id="proyectos" class="projects-section">
@@ -18,14 +21,14 @@ const isSingleProject = PROJECTS.length === 1;
             <h3 data-i18n="projects.subtitle_1">{t('projects.subtitle_1')}</h3>
         </div>   
         <div class={`cards-container ${isSingleProject ? 'single' : ''}`}>
-            {PROJECTS.map(project => (
+            {projects.map(project => (
                 <CardProject classList={['card-project', 'project-element']} {...project} />
             ))}
         </div>
         <div class="colaboration-container">
             <h3 data-i18n="projects.subtitle_2">{t('projects.subtitle_2')}</h3>
             <div class="cards-container colaborate">
-                {PROJECTSC.map(project => (
+                {collaborationProjects.map(project => (
                     <CardColaboracion classlist={['card-project', 'project-element']} {...project} />
                 ))}
             </div>
