+++ src/data/experience.js
@@ -1,34 +1,78 @@
-export const EXPERIENCE = [
-    {
+export const EXPERIENCE = {
+    es: [
+      {
+        id: 'job1',
         title: 'Freelance - Colaboración en Sitio Juana Manso',
         position: "Diseñadora y Frontend Developer",
         url: "https://www.juanamanso.org/",
         startDate: 'Agosto 2024',
         endDate: "",
         description: 'Rediseño del Home en WordPress utilizando Elementor, implementando un slider interactivo y nuevas secciones. Creación de nuevas páginas de contenido y modificación del footer, agregando una nueva sección. Aseguré la responsividad del sitio para una mejor experiencia en dispositivos móviles.',
-    },
-    {
+      },
+      {
+        id: 'job2',
         title: 'Freelance - Colaboración en ILK Abogados Argentina',
         position: "Frontend Developer",
         url: "https://ilk.com.ar/",
         startDate: 'Marzo 2024',
         endDate: "",
         description: 'Desarrollo de un slider con video de fondo en la sección Hero de la homepage, utilizando WordPress y el constructor Divi. Aseguré que el diseño fuera totalmente responsivo para mejorar la experiencia en dispositivos móviles.',
-    },
-    {
-        title: 'Publissoft - Programming Intern',
-        position: "Programming Intern",
+      },
+      {
+        id: 'job3',
+        title: 'Publissoft - Interna de Programación',
+        position: "Interna de Programación",
         url: "",
         startDate: 'Enero 2024',
         endDate: "Actualidad",
         description: 'Gestiono la integración de contenido y la publicación de artículos en diversos CMS como WordPress, Shopify, Wix, Prestashop y CPanel. Mi trabajo incluye la creación de páginas web, la optimización de contenido para mejorar la visibilidad y la experiencia del usuario, y la colaboración con equipos multidisciplinarios para asegurar que el contenido esté alineado con las estrategias y objetivos del negocio. Mantengo los sitios web actualizados y relevantes, adaptando el contenido a las necesidades cambiantes.',
-    },
-    {
+      },
+      {
+        id: 'job4',
         title: 'Freelance - Colaboración de Maquetación',
         position: "Frontend Developer",
         url: "https://www.goigro.com/es/",
         startDate: 'Octubre 2023',  
         endDate: "", 
         description: 'Desarrollo de un banner interactivo en el CMS HubSpot utilizando HTML y CSS. El banner permite al cliente cambiar las imágenes y el texto de cada ítem, y al hacer hover, cambia dinámicamente tanto el texto como la imagen.',
-    },
-];
+      },
+    ],
+    en: [
+      {
+        id: 'job1',
+        title: 'Freelance - Collaboration on Juana Manso Site',
+        position: "Designer and Frontend Developer",
+        url: "https://www.juanamanso.org/",
+        startDate: 'August 2024',
+        endDate: "",
+        description: 'Redesign of the Home in WordPress using Elementor, implementing an interactive slider and new sections. Created new content pages and modified the footer by adding a new section. Ensured site responsiveness for a better experience on mobile devices.',
+      },
+      {
+        id: 'job2',
+        title: 'Freelance - Collaboration on ILK Abogados Argentina',
+        position: "Frontend Developer",
+        url: "https://ilk.com.ar/",
+        startDate: 'March 2024',
+        endDate: "",
+        description: 'Developed a slider with a video background in the Hero section of the homepage, using WordPress and the Divi builder. Ensured the design was fully responsive to improve the mobile experience.',
+      },
+      {
+        id: 'job3',
+        title: 'Publissoft - Programming Intern',
+        position: "Programming Intern",
+        url: "",
+        startDate: 'January 2024',
+        endDate: "Present",
+        description: 'Managed content integration and publication of articles on various CMS platforms such as WordPress, Shopify, Wix, Prestashop, and CPanel. My work includes creating web pages, optimizing content to improve visibility and user experience, and collaborating with cross-functional teams to ensure content aligns with business strategies and objectives. Kept websites updated and relevant by adapting content to evolving needs.',
+      },
+      {
+        id: 'job4',
+        title: 'Freelance - Layout Collaboration',
+        position: "Frontend Developer",
+        url: "https://www.goigro.com/es/",
+        startDate: 'October 2023',
+        endDate: "", 
+        description: 'Developed an interactive banner in the HubSpot CMS using HTML and CSS. The banner allows the client to change the images and text of each item, and on hover, both the text and the image dynamically change.',
+      },
+    ],
+  };
