+++ src/data/projects_colaborate.js
@@ -1,20 +1,48 @@
-export const PROJECTSC = [
-    {
+export const PROJECTSC = {
+    es: [
+      {
         image: '/cardsImg/juanaManso/juanaManso.png',
         title: 'Juana Manso rediseño',
-        description: 'Se realizo el rediseño de la página home y otras páginas en Wordpress, agregando nuevas secciones que estilizan el sitio y dan una mejor accesibilidad al usuario. También agregue navegabilidad al navbar y las nuevas páginas.',
-        link: 'https://www.juanamanso.org/#',        
-    },
-    {
+        description: 'Se realizó el rediseño de la página home y otras páginas en WordPress, agregando nuevas secciones que estilizan el sitio y mejoran la accesibilidad para el usuario. También agregué navegabilidad al navbar y las nuevas páginas.',
+        view_more:'ver más',
+        link: 'https://www.juanamanso.org/#',
+      },
+      {
         image: '/cardsImg/ilkAbogados/IlkAbogadosHeader.jpg',
         title: 'Hero Slider de ILK Abogados Argentina',
-        description: 'Hero slider con video de fondo realizado en Wordpress con el constructor Divi, totalmente responsivo.',
-        link: 'https://ilk.com.ar/',        
-    },
-    {
+        description: 'Hero slider con video de fondo realizado en WordPress con el constructor Divi, totalmente responsivo.',
+        view_more:'ver más',
+        link: 'https://ilk.com.ar/',
+      },
+      {
         image: '/cardsImg/Igro/Igro-Banner.png',
         title: 'Banner del Hero de Igro',
-        description: 'Banner interactivo realizado en el CMS HubSpot con HTML y CSS. Permite al cliente cambiar las imágenes y el texto de cada ítem y al hacer hover, cambia dinámicamente tanto el texto como la imágen.',
-        link: 'https://www.goigro.com/es/',        
-    },
-];
+        description: 'Banner interactivo realizado en el CMS HubSpot con HTML y CSS. Permite al cliente cambiar las imágenes y el texto de cada ítem, y al hacer hover, cambia dinámicamente tanto el texto como la imagen.',
+        view_more:'ver más',
+        link: 'https://www.goigro.com/es/',
+      },
+    ],
+    en: [
+      {
+        image: '/cardsImg/juanaManso/juanaManso.png',
+        title: 'Juana Manso Redesign',
+        description: 'The redesign of the home page and other pages in WordPress, adding new sections that stylize the site and improve user accessibility. I also added navigation to the navbar and the new pages.',
+        view_more:'view more',
+        link: 'https://www.juanamanso.org/#',
+      },
+      {
+        image: '/cardsImg/ilkAbogados/IlkAbogadosHeader.jpg',
+        title: 'Hero Slider for ILK Abogados Argentina',
+        description: 'Hero slider with a background video made in WordPress using the Divi builder, fully responsive.',
+        view_more:'view more',
+        link: 'https://ilk.com.ar/',
+      },
+      {
+        image: '/cardsImg/Igro/Igro-Banner.png',
+        title: 'Hero Banner for Igro',
+        description: 'Interactive banner created in HubSpot CMS using HTML and CSS. It allows the client to change the images and text of each item, and on hover, both the text and image dynamically change.',
+        view_more:'view more',
+        link: 'https://www.goigro.com/es/',
+      },
+    ],
+  };
