--- src/data/projects_colaborate2.js
@@ -1,42 +1,0 @@
-export const PROJECTSC = {
-    es: [
-      {
-        image: '/cardsImg/juanaManso/juanaManso.png',
-        title: 'Juana Manso rediseño',
-        description: 'Se realizó el rediseño de la página home y otras páginas en WordPress, agregando nuevas secciones que estilizan el sitio y mejoran la accesibilidad para el usuario. También agregué navegabilidad al navbar y las nuevas páginas.',
-        link: 'https://www.juanamanso.org/#',
-      },
-      {
-        image: '/cardsImg/ilkAbogados/IlkAbogadosHeader.jpg',
-        title: 'Hero Slider de ILK Abogados Argentina',
-        description: 'Hero slider con video de fondo realizado en WordPress con el constructor Divi, totalmente responsivo.',
-        link: 'https://ilk.com.ar/',
-      },
-      {
-        image: '/cardsImg/Igro/Igro-Banner.png',
-        title: 'Banner del Hero de Igro',
-        description: 'Banner interactivo realizado en el CMS HubSpot con HTML y CSS. Permite al cliente cambiar las imágenes y el texto de cada ítem, y al hacer hover, cambia dinámicamente tanto el texto como la imagen.',
-        link: 'https://www.goigro.com/es/',
-      },
-    ],
-    en: [
-      {
-        image: '/cardsImg/juanaManso/juanaManso.png',
-        title: 'Juana Manso Redesign',
-        description: 'The redesign of the home page and other pages in WordPress, adding new sections that stylize the site and improve user accessibility. I also added navigation to the navbar and the new pages.',
-        link: 'https://www.juanamanso.org/#',
-      },
-      {
-        image: '/cardsImg/ilkAbogados/IlkAbogadosHeader.jpg',
-        title: 'Hero Slider for ILK Abogados Argentina',
-        description: 'Hero slider with a background video made in WordPress using the Divi builder, fully responsive.',
-        link: 'https://ilk.com.ar/',
-      },
-      {
-        image: '/cardsImg/Igro/Igro-Banner.png',
-        title: 'Hero Banner for Igro',
-        description: 'Interactive banner created in HubSpot CMS using HTML and CSS. It allows the client to change the images and text of each item, and on hover, both the text and image dynamically change.',
-        link: 'https://www.goigro.com/es/',
-      },
-    ],
-  };
