--- src/data/projects_detail.js
@@ -1,26 +1,0 @@
-export const PROJECTS = [
-    {
-        slug: 'avocado', 
-        fecha: 'junio 2024',
-        image: '/cardsImg/avocado/avocadoFeed.jpg',
-        image2: '/cardsImg/avocado/logo-palta.png',
-        imageD: '/cardsImg/avocado/avocadoFC.JPG',
-        title: 'Avocado',
-        description: 'Red Social de Recetas realizado para la Tecnicatura en Desarrollo Web y Aplicaciones Digitales.',
-        descriptionExtend: 'Avocado es una aplicación web diseñada para que los usuarios registrados creen y compartan recetas con la comunidad y puedan adquirir suscripciones de servicios. Esto permite explorar nuevos sabores y fomentar la creatividad culinaria. Los usuarios pueden descubrir sugerencias en el feed de la aplicación y guardar sus recetas favoritas para un acceso conveniente. Además, pueden aprovechar las funciones de búsqueda y filtrado para resultados específicos en sus exploraciones culinarias. Nuestro enfoque estaba en crear una plataforma atractiva y funcional, trabajamos en estrecha colaboración para garantizar que el sitio web refleje la idea que teníamos en mente.',
-        rol:'Desarrollé una interfaz de usuario dinámica y altamente interactiva utilizando Angular, creando componentes personalizados para mejorar tanto el atractivo visual como la funcionalidad. Además, me aseguré de que la página fuera totalmente adaptable a varios dispositivos, garantizando así una experiencia óptima para todos los usuarios.',
-        tecnologias: ['Angular','TypeScript', 'HTML5', 'CSS3','Node.js', 'Express', 'MySQL']
-    },
-/*     {
-        slug: 'portfolio',  
-        fecha: 'agosto 2024',
-        image: '/cardsImg/portfolio/portfolio.jpg',
-        image2: '/cardsImg/portfolio/logo-portfolio.png',
-        imageD: '/cardsImg/portfolio/portfolioFC.jpg',
-        title: 'Personal Portfolio',
-        description: 'Portfolio personal, aquí encontraras mis proyectos.',
-        descriptionExtend:'Mi portfolio como Developer está diseñado para mostrar mis habilidades y proyectos a través de una interfaz atractiva y funcional. El sitio permite a los visitantes explorar mis trabajos, incluyendo ejemplos de desarrollo web, animaciones y diseño responsivo. Los usuarios pueden navegar por diferentes secciones que destacan mis colaboraciones en proyectos reales, así como mi experiencia en la creación de interfaces intuitivas y dinámicas. Mi enfoque principal fue garantizar que el portafolio no solo reflejara mi capacidad técnica, sino también mi creatividad y atención al detalle, proporcionando una experiencia de usuario fluida y agradable.',
-        rol:'A lo largo del desarrollo de mi portfolio, encabezé la implementación de diseño y funcionalidad utilizando Astro y JavaScript, incorporando bibliotecas como GSAP para una interfaz de usuario atractiva y animaciones dinámicas. Me centré en el diseño receptivo, la optimización del rendimiento y la presentación efectiva del contenido para mostrar mis habilidades y proyectos de manera clara y concisa. Además, me aseguré de que la página fuera totalmente adaptable a varios dispositivos, garantizando así una experiencia óptima para todos los usuarios.',
-        tecnologias: ['Astro','JavaScript', 'HTML5', 'CSS3', 'GSAP']
-    },  */
-];
