+++ src/data/projects_details.js
@@ -10,7 +10,8 @@ export const PROJECTS = {
         description: 'Red Social de Recetas realizado para la Tecnicatura en Desarrollo Web y Aplicaciones Digitales.',
         descriptionExtend: 'Avocado es una aplicación web diseñada para que los usuarios registrados creen y compartan recetas con la comunidad y puedan adquirir suscripciones de servicios. Esto permite explorar nuevos sabores y fomentar la creatividad culinaria. Los usuarios pueden descubrir sugerencias en el feed de la aplicación y guardar sus recetas favoritas para un acceso conveniente. Además, pueden aprovechar las funciones de búsqueda y filtrado para resultados específicos en sus exploraciones culinarias. Nuestro enfoque estaba en crear una plataforma atractiva y funcional, trabajamos en estrecha colaboración para garantizar que el sitio web refleje la idea que teníamos en mente.',
         rol: 'Desarrollé una interfaz de usuario dinámica y altamente interactiva utilizando Angular, creando componentes personalizados para mejorar tanto el atractivo visual como la funcionalidad. Además, me aseguré de que la página fuera totalmente adaptable a varios dispositivos, garantizando así una experiencia óptima para todos los usuarios.',
-        tecnologias: ['Angular', 'TypeScript', 'HTML5', 'CSS3', 'Node.js', 'Express', 'MySQL']
+        tecnologias: ['Angular', 'TypeScript', 'HTML5', 'CSS3', 'Node.js', 'Express', 'MySQL'],
+        link: 'ver más',
       },
     ],
     en: [
@@ -24,7 +25,8 @@ export const PROJECTS = {
         description: 'Recipe Social Network developed for the Web Development and Digital Applications Program.',
         descriptionExtend: 'Avocado is a web application designed for registered users to create and share recipes with the community, as well as acquire subscription services. This allows users to explore new flavors and foster culinary creativity. Users can discover suggestions in the app’s feed and save their favorite recipes for convenient access. They can also take advantage of search and filtering features for specific results in their culinary explorations. Our focus was on creating an attractive and functional platform, and we worked closely to ensure that the website reflected the idea we had in mind.',
         rol: 'I developed a dynamic and highly interactive user interface using Angular, creating custom components to enhance both visual appeal and functionality. I also ensured that the page was fully responsive across various devices, thus guaranteeing an optimal experience for all users.',
-        tecnologias: ['Angular', 'TypeScript', 'HTML5', 'CSS3', 'Node.js', 'Express', 'MySQL']
+        tecnologias: ['Angular', 'TypeScript', 'HTML5', 'CSS3', 'Node.js', 'Express', 'MySQL'],
+        link: 'view more',
       },
     ]
   };
