+++ src/i18n/utils.js
@@ -1,4 +1,6 @@
 import { translations } from '@/i18n/translations';
+import { EXPERIENCE } from '@/data/experience';
+
 
 
 const defaultLang = 'es';
@@ -25,25 +27,47 @@ export function updateLanguage(lang) {
     localStorage.setItem('lang', lang);
     
     const elements = document.querySelectorAll('[data-i18n]');
-    console.log('Elementos con data-i18n:', elements); // Verifica que encuentra los elementos a traducir
+    console.log('Elementos con data-i18n:', elements);
     
     elements.forEach(el => {
       const key = el.getAttribute('data-i18n');
       const translation = useTranslations(lang)(key);
       
-      // Busca solo el nodo de texto dentro del elemento y actualízalo
       const textNode = Array.from(el.childNodes).find(node => node.nodeType === Node.TEXT_NODE);
       
       if (textNode) {
         textNode.nodeValue = translation;
       } else {
-        // Si no hay un nodo de texto, lo creamos
         el.insertBefore(document.createTextNode(translation), el.firstChild);
       }
 
       console.log(`Actualizando ${key} con la traducción:`, translation);
     });
+
+    // Update experience items
+    updateExperienceItems(lang);
   }
+}
+
+function updateExperienceItems(lang) {
+  const experienceItems = document.querySelectorAll('.experience-item');
+  const translations = EXPERIENCE[lang] || EXPERIENCE[defaultLang];
+
+  experienceItems.forEach((item, index) => {
+    const experience = translations[index];
+    if (experience) {
+      item.querySelector('h3').textContent = experience.title;
+      item.querySelector('h4').textContent = experience.position;
+      item.querySelector('time').textContent = experience.endDate 
+        ? `${experience.startDate} - ${experience.endDate}`
+        : experience.startDate;
+      item.querySelector('p').textContent = experience.description;
+      const link = item.querySelector('a');
+      if (link) {
+        link.href = experience.url;
+      }
+    }
+  });
 }
 
 export function getCurrentLang() {
